server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
bcrypt==4.1.2
email-validator==2.1.0
greenlet==3.0.3
asyncpg==0.29.0
pgvector==0.2.5
streamlit==1.37.1
pandas==2.2.2
apscheduler==3.10.4
